Fixed stormwind guard equipment change, but it is not proper way... Fixed compile wraning in ZulAman script Forgotten change in past commit, in Icecrown Citadel script

+++ scripts/RaidScripts/Raid_ZulAman.cpp
@@ -331,7 +331,7 @@ class HalazziAI : public MoonScriptBossAI
 		int32 mTotemTimer;
 		int32 CurrentHealth;
 		int32 MaxHealth;
-		int SplitCount;
+		uint8 SplitCount;
 };
 
 class LynxSpiritAI : public MoonScriptCreatureAI
